requirements.txt
Flask
requests
gunicorn
openai-whisper
torch
torchaudio
coqui-tts